// src/app/shared/model/employee/employee-response.ts
export class EmployeeResponse {
  id: number;
  userName: string;
  password: string;
  email: string;
  firstName: string;
  lastName: string;
  phone: string;
  address: string;
  image: string;
  role: string;
  createdDate: string;
  sex: string;
  enabled: boolean;
  status: boolean;
  dateOfBirth: string;
  id_creator: number;
  id_homeStay: number;
  homeStayName: string;
}